Remove deepchecks dependency and reorganize related tests.
Removed all references to deepchecks and its associated components, fixtures, and tests. Introduced new test files and restructured certain modules to streamline functionality and adhere to the updated project organization. Updated dev dependencies and other dependencies to ensure compatibility with these changes.
