+++ tabular_orchestrated/mljar/mljar_deepchecks.py
@@ -1,13 +1,13 @@
-import dataclasses
-
-from tabular_orchestrated.dc.dc_model import DCFullComp, DCModelComp
-
-
-@dataclasses.dataclass
-class MljarDCModelComp(DCModelComp):
-    extra_packages = ["deepchecks", "mljar"]
-
-
-@dataclasses.dataclass
-class MljarDCFullComp(DCFullComp):
-    extra_packages = ["deepchecks", "mljar"]
+# import dataclasses
+#
+# from tabular_orchestrated.dc.dc_model import DCFullComp, DCModelComp
+#
+#
+# @dataclasses.dataclass
+# class MljarDCModelComp(DCModelComp):
+#     extra_packages = ["deepchecks", "mljar"]
+#
+#
+# @dataclasses.dataclass
+# class MljarDCFullComp(DCFullComp):
+#     extra_packages = ["deepchecks", "mljar"]
